fix segfault on 64bit systems

+++ src/miscc.c
@@ -597,7 +597,7 @@ int Z_moveobj(obj_t *p) {
 void Z_splash(obj_t *p,int n) {
   Z_sound(bulsnd[0],128);
   DOT_water(p->x,p->y-p->h/2,p->xv+p->vx,p->yv+p->vy,n,
-	(int)walp[wfront]-1);
+	(intptr_t)walp[wfront]-1);
 }
 
 void Z_calc_time(dword t,word *h,word *m,word *s)
